+++ Entities/AHuman.cpp
@@ -57,6 +57,7 @@ void AHuman::Clear()
     m_StrideSound = nullptr;
     m_ArmsState = WEAPON_READY;
     m_MoveState = STAND;
+	m_LimbPushForcesAndCollisionsDisabled = false;
     m_ProneState = NOTPRONE;
     m_ProneTimer.Reset();
     for (int i = 0; i < MOVEMENTSTATECOUNT; ++i) {
@@ -214,6 +215,8 @@ int AHuman::Create(const AHuman &reference) {
     m_ArmsState = reference.m_ArmsState;
     m_MoveState = reference.m_MoveState;
     m_ProneState = reference.m_ProneState;
+
+	m_LimbPushForcesAndCollisionsDisabled = reference.m_LimbPushForcesAndCollisionsDisabled;
 
     for (int i = 0; i < MOVEMENTSTATECOUNT; ++i) {
         m_Paths[FGROUND][i].Create(reference.m_Paths[FGROUND][i]);
@@ -3546,7 +3549,7 @@ void AHuman::Update()
     ///////////////////////////////////////////////////
     // Travel the limb AtomGroup:s
 
-    if (m_Status == STABLE && m_MoveState != NOMOVE)
+	if (m_Status == STABLE && !m_LimbPushForcesAndCollisionsDisabled && m_MoveState != NOMOVE)
     {
         // This exists to support disabling foot collisions if the limbpath has that flag set.
         if ((m_pFGFootGroup->GetAtomCount() == 0 && m_BackupFGFootGroup->GetAtomCount() > 0) != m_Paths[FGROUND][m_MoveState].FootCollisionsShouldBeDisabled()) {